Factor out asset loading, inventory, and UI into separate modules
